my auth system

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -12,6 +12,8 @@ import com.example.demo.util.Ut;
 import com.example.demo.vo.Article;
 import com.example.demo.vo.ResultData;
 
+import jakarta.servlet.http.HttpSession;
+
 @Controller
 public class UsrArticleController {
 
@@ -26,7 +28,7 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/getArticle")
 	@ResponseBody
-	public ResultData getArticleAction(int id) {
+	public ResultData<Article> getArticleAction(int id) {
 		Article article = articleService.getArticle(id);
 
 		if (article == null) {
@@ -38,22 +40,24 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/getArticles")
 	@ResponseBody
-	public ResultData getArticles() {
+	public ResultData<List<Article>> getArticles() {
 		List<Article> articles = articleService.getArticles();
 		return ResultData.from("S-1", "Article List", articles);
 	}
 
 	@RequestMapping("/usr/article/doWrite")
 	@ResponseBody
-	public ResultData doWrite(String title, String body) {
+	public ResultData<Article> doWrite(String title, String body, HttpSession session) {
 		if (Ut.isNullOrEmpty(title)) {
 			return ResultData.from("F-1", "제목을 입력해주세요");
 		}
 		if (Ut.isNullOrEmpty(body)) {
 			return ResultData.from("F-2", "내용을 입력해주세요");
 		}
+		
+		int writer = Integer.valueOf(session.getAttribute("loginedMemberId").toString());
 
-		ResultData writeArticleRd = articleService.writeArticle(title, body);
+		ResultData<Integer> writeArticleRd = articleService.writeArticle(title, body, writer);
 
 		int id = (int) writeArticleRd.getData1();
 
@@ -62,38 +66,42 @@ public class UsrArticleController {
 		return ResultData.from(writeArticleRd.getResultCode(), writeArticleRd.getMsg(), article);
 	}
 
-
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(int id, String title, String body) {
-		System.err.println("title : " + title);
-		System.err.println("body : " + body);
+	public ResultData<Integer> doModify(int id, String title, String body, HttpSession session) {
 		Article article = articleService.getArticle(id);
 
 		if (article == null) {
-			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다", id));
+			return ResultData.from("F-1", Ut.f("%d번 글은 존재하지 않습니다", id), id);
+		}
+		
+		
+		if(Integer.valueOf(String.valueOf(session.getAttribute("loginedMemberId")))!=article.getWriter()) {
+			return ResultData.from("F-2", "해당글의 작성자만 수정 기능을 이용할 수 있습니다.");
 		}
 
 		articleService.modifyArticle(id, title, body);
-		Article findArticle = articleService.getArticle(id);
 
-		return ResultData.from("S-1", Ut.f("%d번 게시물이 수정되었습니다.", id), findArticle);
+		return ResultData.from("S-1", Ut.f("%d번 글을 수정했습니다", id), id);
 	}
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData<Integer> doDelete(int id, HttpSession session) {
 
 		Article article = articleService.getArticle(id);
 
 		if (article == null) {
-			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다", id));
+			return ResultData.from("F-1", Ut.f("%d번 글은 존재하지 않습니다", id), id);
+		}
+		
+		if(Integer.valueOf(String.valueOf(session.getAttribute("loginedMemberId")))!=article.getWriter()) {
+			return ResultData.from("F-2", "해당글의 작성자만 삭제 기능을 이용할 수 있습니다.");
 		}
 
 		articleService.deleteArticle(id);
 
-		return ResultData.from("S-1", Ut.f("%d번 게시물이 삭제되었습니다.", id));
+		return ResultData.from("S-1", Ut.f("%d번 글이 삭제 되었습니다", id), id);
 	}
-
 
 }
